+++ lib/commands/flutter.dart
@@ -1,7 +1,8 @@
 import 'package:args/command_runner.dart';
 import 'package:fvm/constants.dart';
+import 'package:fvm/flutter/flutter_helpers.dart';
 import 'package:fvm/flutter/flutter_tools.dart';
-import 'package:fvm/utils/helpers.dart';
+
 import 'package:args/args.dart';
 
 /// Proxies Flutter Commands
