+++ lib/flutter/flutter_helpers.dart
@@ -1,8 +1,10 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:fvm/constants.dart';
 import 'package:fvm/exceptions.dart';
-
+import 'package:fvm/utils/project_config.dart';
+import 'package:path/path.dart' as path;
 import 'flutter_releases.dart';
 
 /// Returns true if it's a valid Flutter version number
@@ -30,4 +32,31 @@ Future<String> inferFlutterVersion(String version) async {
 /// Returns true if it's a valid Flutter channel
 bool isFlutterChannel(String channel) {
   return kFlutterChannels.contains(channel);
+}
+
+/// Check if it is the current version.
+bool isCurrentVersion(String version) {
+  final configVersion = getConfigFlutterVersion();
+  return version == configVersion;
+}
+
+/// Checks if its global version
+bool isGlobalVersion(String version) {
+  if (!kDefaultFlutterLink.existsSync()) return false;
+
+  final globalVersion = path.basename(kDefaultFlutterLink.targetSync());
+
+  return globalVersion == version;
+}
+
+String getFlutterSdkExec({String version}) {
+  return path.join(getFlutterSdkPath(version: version), 'bin',
+      Platform.isWindows ? 'flutter.bat' : 'flutter');
+}
+
+/// The Flutter SDK Path referenced on FVM
+String getFlutterSdkPath({String version}) {
+  var sdkVersion = version;
+  sdkVersion ??= getConfigFlutterVersion();
+  return path.join(kVersionsDir.path, sdkVersion);
 }
