+++ lib/flutter/flutter_tools.dart
@@ -4,6 +4,7 @@ import 'package:fvm/exceptions.dart';
 
 import 'package:fvm/utils/git.dart';
 
+import 'package:fvm/flutter/flutter_helpers.dart';
 import 'package:fvm/utils/helpers.dart';
 import 'package:fvm/utils/logger.dart';
 import 'package:fvm/utils/pretty_print.dart';
