+++ lib/utils/project_config.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:fvm/constants.dart';
 import 'package:fvm/exceptions.dart';
 import 'package:fvm/utils/helpers.dart';
+import 'package:fvm/flutter/flutter_helpers.dart';
 import 'package:fvm/utils/pretty_print.dart';
 
 class ProjectConfig {
